requirements.txt
streamlit==1.33.0
google-cloud-firestore
google.generativeai
google-auth
pandas==2.1.2